[Finishes #168903089] added middleware for validating endpoints
