+++ .travis.yml
@@ -1,5 +1,10 @@
 language: node_js
+
 node_js:
   - "10.16.3"
-script: npm run coverage
-after_success: npm run coveralls
+  
+script:
+  - npm test
+
+after_success:
+  - npm coverage
