+++ package.json
@@ -4,7 +4,7 @@
   "description": "andela project",
   "main": "index.js",
   "scripts": {
-    "test": "node node_modules/.bin/mocha",
+    "test": "mocha --require @babel/register server/**/**/**tests.js --exit",
     "build": "babel server -d build",
     "start": "babel-node server/index.js",
     "dev": "nodemon --exec babel-node server/index.js",
