+++ server/index.js
@@ -3,7 +3,7 @@ import routes from './routes/routes';
 import body_parser from 'body-parser';
 
 const app = express()
-const PORT = process.env.PORT || 8000;
+const PORT = process.env.PORT || 3000;
 
 app.use(express.json())
 app.use(body_parser.urlencoded({extended: true}));
