+++ server/middleware/loginValidator.js
@@ -1,0 +1,16 @@
+import Joi from '@hapi/joi';
+
+export default (req, res, next)=>{
+    const schema = {
+        email: Joi.string().email().required(),
+        password: Joi.string().min(6).required(),
+    }
+    const result = Joi.validate(req.body, schema);
+    if (result.error) {
+        return res.status(400).send({
+            "status": 400,
+            "error": result.error.details[0].message
+        });
+    }
+    next()
+}
